Moved views to their own files from both FoodTruckProfileView and FoodTruckEditView so that it's easier to navigate and fix code but also so that the compilator don't get worked too hard. Also indented every file and changed every primary class/struct names to their respective filename
