+++ Components/FoodTruckService.swift
@@ -12,7 +12,7 @@ import FirebaseFirestoreSwift
 class FoodTruckService {
     private var db = Firestore.firestore()
     
-    func fetchFoodTruck(by truckId: String, completion: @escaping (FoodTruck?) -> Void) {
+    func fetchFoodTruck(by truckId: String, completion: @escaping (FoodTruckModel?) -> Void) {
         db.collection("foodTrucks").document(truckId).getDocument { (document, error) in
             if let document = document, document.exists {
                 do {
@@ -31,7 +31,7 @@ class FoodTruckService {
                         data["locationPeriod"] = ""
                     }
                     let jsonData = try JSONSerialization.data(withJSONObject: data, options: [])
-                    let foodTruck = try JSONDecoder().decode(FoodTruck.self, from: jsonData)
+                    let foodTruck = try JSONDecoder().decode(FoodTruckModel.self, from: jsonData)
                     completion(foodTruck)
                 } catch {
                     print("Error decoding food truck: \(error)")
@@ -44,7 +44,7 @@ class FoodTruckService {
         }
     }
     
-    func fetchFoodTrucks(completion: @escaping ([FoodTruck]) -> Void) {
+    func fetchFoodTrucks(completion: @escaping ([FoodTruckModel]) -> Void) {
         db.collection("foodTrucks").getDocuments { (querySnapshot, error) in
             if let error = error {
                 print("Error getting documents: \(error)")
@@ -52,7 +52,7 @@ class FoodTruckService {
                 return
             }
             
-            let foodTrucks = querySnapshot?.documents.compactMap { document -> FoodTruck? in
+            let foodTrucks = querySnapshot?.documents.compactMap { document -> FoodTruckModel? in
                 do {
                     var data = document.data()
                     if data["dailyDeals"] == nil {
@@ -68,7 +68,7 @@ class FoodTruckService {
                         data["locationPeriod"] = ""
                     }
                     let jsonData = try JSONSerialization.data(withJSONObject: data, options: [])
-                    let foodTruck = try JSONDecoder().decode(FoodTruck.self, from: jsonData)
+                    let foodTruck = try JSONDecoder().decode(FoodTruckModel.self, from: jsonData)
                     return foodTruck
                 } catch {
                     print("Error decoding food truck: \(error)")
@@ -79,7 +79,7 @@ class FoodTruckService {
         }
     }
     
-    func saveFoodTruck(_ foodTruck: FoodTruck, completion: @escaping (Error?) -> Void) {
+    func saveFoodTruck(_ foodTruck: FoodTruckModel, completion: @escaping (Error?) -> Void) {
         do {
             try db.collection("foodTrucks").document(foodTruck.id).setData(from: foodTruck) { error in
                 completion(error)
