+++ Models/ReviewModel.swift
@@ -7,14 +7,14 @@
 
 import Foundation
 
-struct Review: Identifiable, Codable {
+struct ReviewModel: Identifiable, Codable {
     var id: String = UUID().uuidString
     var userId: String
     var userName: String
     var text: String
     var rating: Double
     var foodTruckName: String?
-
+    
     init(id: String = UUID().uuidString, userId: String, userName: String, text: String, rating: Double, foodTruckName: String? = nil) {
         self.id = id
         self.userId = userId
@@ -24,4 +24,3 @@ struct Review: Identifiable, Codable {
         self.foodTruckName = foodTruckName
     }
 }
-
