+++ Views/GoogleSignInButtonView.swift
@@ -99,7 +99,7 @@ struct GoogleSignInButton: UIViewRepresentable {
                         let email = user.email ?? "No Email"
                         let username = user.displayName ?? "No Name"
                         let userType = 1
-                        UserManager.shared.userType = 1
+                        UserManagerModel.shared.userType = 1
                         
                         self.saveUserProfile(uid: uid, username: username, email: email, userType: userType)
                     }
