--- Views/LocationMapView.swift
@@ -1,25 +1,0 @@
-//
-//  LocationMapView.swift
-//  FoodTruckFinder
-//
-//  Created by Oskar Lövstrand on 2024-05-19.
-//
-
-import SwiftUI
-import MapKit
-
-struct LocationMapView: UIViewRepresentable {
-    var coordinate: CLLocationCoordinate2D
-    
-    func makeUIView(context: Context) -> MKMapView {
-        MKMapView(frame: .zero)
-    }
-    
-    func updateUIView(_ view: MKMapView, context: Context) {
-        let region = MKCoordinateRegion(
-            center: coordinate,
-            span: MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
-        )
-        view.setRegion(region, animated: true)
-    }
-}
